Add pagination test for products without description
